Sets up linting, type checking, and testing rules and enforces them in new Github workflow (#17)
Closes #13 #14 #15 

The diffs in the regular python files are just formatting changes
enforced by ruff. In the `__init__.py` files, adding the redundant alias
is done to communicate that these are purposefully "unused" imports that
are being used to expose a public interface to other modules (see
[unused-import
(F401)](https://docs.astral.sh/ruff/rules/unused-import/#unused-import-f401)).

As a starting point, I'm setting the minimum test coverage to 90%.