Creates Strava transfer service class and adds configuration to base transfer service class (#39)
Closes #18 

**State as arg**: When adding `pardner` to `pardner-site`, I realized
it's necessary to instantiate a transfer service with an existing
`state` variable defined, so I added that as an argument to the
constructor.

**Configuring scope**: `requests_oauthlib` always sends requests with
scopes separated with a space, but Strava (for example) specifies the
scopes need to be comma-separated.

**Adding default logic to base class**: The `fetch_token` and
`authorization_url` methods will almost always have the same
functionality regardless of service, so I filled them out in the base
class as well.

**Tests**: extracted common test fixtures and moved some tests to base
transfer service test file.