GroupMe OAuth and data fetching methods, new Verticals, and some refactoring (#55)
Apologies for the long PR! More than half of it is just tests and method
descriptions.

Closes #41 #54 

**GroupMe**
- Implements OAuth flow. Note that GroupMe doesn't require the
background server (the one running `pardner`) to make a request to its
authorization server to obtain the access token. That happens in a
single request from the client.
- Implements methods for fetching **new verticals**: conversations,
chatbots, and blocked users.

**Base Service class configuration and helper methods**
- `client_secret` is optional now since GroupMe doesn't use it.
- `_get_resource_from_path` for passing in just the unique part of the
API endpoint path (so `'users'` in `'https://api.tumblr.com/v2/users'`,
for example).
- `_build_resource_url` for getting the full API endpoint URL from the
unique part described in the bullet point above.
- Updated Strava and Tumblr classes to use new helper.