Defines custom models for verticals (#65)
Closes #60 

This PR defines the models for the different verticals we support in
`pardner` at the moment. I opted to use
[Pydantic](https://docs.pydantic.dev/latest/) to define these models
since it's flexible, leverages a lot of typing concepts already built
into Python, and is easy to serialize/deserialize.

As a reminder, we don't have plans to store any data in `pardner`,
that's the responsibility of the user of the library.

Next up:
- #61
- #62
- #63
- #64