Uses subclasses of `BaseVertical` rather than the `Vertical` enum to identify a vertical (#66)
Closes #61 

So any time we were using `Vertical.SubVerticalName` we now use
`SubVerticalName`, where `SubVerticalName` is a subclass of
`BaseVertical`. Note that I do mean subclass and not instance of a
subclass since we're using the classes themselves as identifiers of
verticals. In the future, methods will be returning instances of those
same subclasses.

I also adjusted the names of the vertical fetching methods to follow the
structure: `fetch_<vertical_name>_vertical` for consistency and to avoid
having to keep track of the plural names like before.

### Misc
Anywhere where an argument `arg` to a function/method could have one or
more items, like in the constructor of the exception classes, I make
these into a `*arg` instead so we don't have to special case passing in
different number of arguments.