Strava fetch methods return model objects rather than `Any` (#68)
Closes #64 

I decided to have the fetch methods return values with the following
type signature: `tuple[list[<Vertical> | None], Any]`. The first element
in the tuple is the list of model objects that were parsed using the
response from the API request. It's also possibly `None` since there may
be a parsing failure. The second element is the raw response payload,
which we return in case the caller wants more data than we provide in
the model definition.

In this PR I wrote two Strava public methods:
`fetch_social_posting_vertical` and `fetch_physical_activity_vertical`.

I also set the `BaseVertical.id` field to be self-generating.